% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_details_section.R
\name{add_file_details_section}
\alias{add_file_details_section}
\title{Add a File Details section in README
If the README.md in the present working directory does not contain a "# File Details" heading, it is added in the end of it.}
\usage{
add_file_details_section()
}
\description{
Add a File Details section in README
If the README.md in the present working directory does not contain a "# File Details" heading, it is added in the end of it.
}
