% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FileDetails.R
\name{add_FileDetails}
\alias{add_FileDetails}
\title{Create a Readi}
\usage{
add_FileDetails(sources, output_file, summary, ..., disable_rstudioapi = FALSE)
}
\description{
A Readi is an S4 object that contains a dataframe of File Details elements pertaining to a specific csv or Excel file os well as the entire README contents of the working directory. This function makes a new Readi S4 Class object using the given parameters.
This is first achived by:
1. Making a README.md in the working directory if one does not exist, followed by:
2. Adding a File Details section in the README if that does not exist.
}
