% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FileDetails_to_csv.R
\name{add_FileDetails_to_csv}
\alias{add_FileDetails_to_csv}
\title{Add Readi object to a csv file by converting it into Excel}
\usage{
add_FileDetails_to_csv(FileDetails_obj, path_to_csv, commit = TRUE)
}
\description{
Add Readi object to a csv file by converting it into Excel
}
