% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FileDetails_to_Excel.R
\name{add_FileDetails_to_Excel}
\alias{add_FileDetails_to_Excel}
\title{Update an Excel file with a File Details tab}
\usage{
add_FileDetails_to_Excel(FileDetails_obj, path_to_excel)
}
\description{
Update an Excel file with a File Details tab
}
