% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_README_entry.R
\name{add_README_entry}
\alias{add_README_entry}
\title{Add a an entry to README.md section}
\usage{
add_README_entry(section_name, ...)
}
\arguments{
\item{section_name}{name of the new section, excluding the "# " prefix}

\item{...}{comma-separated character lines that will be appended to the README}
}
\description{
Add a an entry to README.md section
}
