% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_README_section.R
\name{add_README_section}
\alias{add_README_section}
\title{Add a section to README.md if it does not already exist
If the README.md in the present working directory does not contain a section with the heading in the format of "# {section_name}", it is added to it at the bottom.}
\usage{
add_README_section(section_name)
}
\arguments{
\item{section_name}{name of the new section, excluding the "# " prefix}
}
\description{
Add a section to README.md if it does not already exist
If the README.md in the present working directory does not contain a section with the heading in the format of "# {section_name}", it is added to it at the bottom.
}
