% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply}
\alias{apply}
\title{Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.}
\usage{
apply(file, sources, summary, path_to_original_output_file, notes)
}
\description{
Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.
}
