% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_readi.R
\name{apply_readi}
\alias{apply_readi}
\title{Takes a Readi Class object and updates the README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.}
\usage{
apply_readi(file, readi_object)
}
\description{
Takes a Readi Class object and updates the README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.
}
