% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_to_csv.R
\name{attach_to_csv}
\alias{attach_to_csv}
\title{Add Readi object to a csv file by converting it into Excel}
\usage{
attach_to_csv(readi_object, path_to_csv)
}
\description{
Add Readi object to a csv file by converting it into Excel
}
