% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_readi_object.R
\name{cast_readi_object}
\alias{cast_readi_object}
\title{Takes a Readi Object and makes the changes to the provided arguments}
\usage{
cast_readi_object(file, readi_object)
}
\description{
Takes a Readi Object and makes the changes to the provided arguments
}
