% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_README_section.R
\name{clear_README_section}
\alias{clear_README_section}
\title{Clear the contents of a README section}
\usage{
clear_README_section(section_name)
}
\description{
Clear all the lines in a section, but keep the section_heading
}
