% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_README.md.R
\name{create_README.md}
\alias{create_README.md}
\title{Write a README.md in the working directory if it doesn't exist}
\usage{
create_README.md()
}
\description{
Write a README.md in the working directory if it doesn't exist
}
