% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_README_section.R
\name{delete_README_section}
\alias{delete_README_section}
\title{Delete the contents of a README section}
\usage{
delete_README_section(section_name)
}
\description{
Delete an entire section in a README
}
