% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_FileDetails_class.R
\docType{class}
\name{FileDetails-class}
\alias{FileDetails-class}
\alias{FileDetails}
\title{Set FileDetails S4 Class
A FileDetails object contains the contents of a README File Details section and a dataframe on the same File Details pertaining to a specific Excel or csv file written out of R.}
\description{
Set FileDetails S4 Class
A FileDetails object contains the contents of a README File Details section and a dataframe on the same File Details pertaining to a specific Excel or csv file written out of R.
}
\section{Slots}{

\describe{
\item{\code{dataframe}}{dataframe of the Readi arguments specific to a given dataframe output file}

\item{\code{readme_section}}{vector of the README 'File Details' section}
}}

