% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_FileDetailsText_class.R
\docType{class}
\name{FileDetailsText-class}
\alias{FileDetailsText-class}
\alias{FileDetailsText}
\title{Set FileDetailsText S4 Class
A FileDetailsText object contains the text version of File Details destined for README.md or another wordprocessing file.}
\description{
Set FileDetailsText S4 Class
A FileDetailsText object contains the text version of File Details destined for README.md or another wordprocessing file.
}
\section{Slots}{

\describe{
\item{\code{Long}}{File Details dataframe structured as a list}

\item{\code{Wide}}{File Details dataframes structured more like a dataframe}
}}

