% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_readi_object.R
\name{init_readi_object}
\alias{init_readi_object}
\title{Create a Readi object from arguments and an existing README}
\usage{
init_readi_object(
  sources,
  summary,
  path_to_local_r_script,
  github_link_to_r_script,
  path_to_original_output_file,
  notes,
  log_details
)
}
\description{
Create a Readi object from arguments and an existing README
}
