% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_AddOnFileDetailsDF.R
\name{make_AddOnFileDetailsDF}
\alias{make_AddOnFileDetailsDF}
\title{Create a Readi Dataframe}
\usage{
make_AddOnFileDetailsDF(...)
}
\arguments{
\item{disable_rstudioapi}{TRUE if the RStudioAPI should be disabled. Errors can be thrown in Jobs and running scripts from the command line.}
}
\description{
Create a Readi Dataframe
}
