% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Bookmark.R
\name{make_Bookmark}
\alias{make_Bookmark}
\title{Make a Bookmark S4 Class object}
\usage{
make_Bookmark(Trace_obj)
}
\description{
Bookmarks provide the line of code where objects reported in the Trace object
}
