% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FileDetails.R
\name{make_FileDetails}
\alias{make_FileDetails}
\title{Make a FileDetails class object}
\usage{
make_FileDetails(sources, output_file, summary, disable_rstudioapi = FALSE)
}
\description{
Make a FileDetails class object
}
