% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_readi_df.R
\name{make_readi_df}
\alias{make_readi_df}
\title{Create a Readi Dataframe}
\usage{
make_readi_df(..., disable_rstudioapi = FALSE)
}
\arguments{
\item{disable_rstudioapi}{TRUE if the RStudioAPI should be disabled. Errors can be thrown in Jobs and running scripts from the command line.}
}
\description{
Create a Readi Dataframe
}
