% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_readi_readme.R
\name{make_readi_readme}
\alias{make_readi_readme}
\title{Make the File Details section for the README.md in wd}
\usage{
make_readi_readme(readi_df)
}
\description{
Make the File Details section for the README.md in wd
}
