% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_README.R
\name{make_README}
\alias{make_README}
\title{Make a README Class object}
\usage{
make_README(section_name)
}
\arguments{
\item{section_name}{This is the name of the README.md section to perform the split into 3 on. The argument should only be the complete string of the section heading for the README.md. For example for the "# File Details", the section_name would be "File Details". Since the heading can be followed by either carriage returns or spaces depending on the style of markdown, an exact string match is not used when splitting the README.md read_lines vector. Should there be more than 1 match, the first index returned by the grep function is used to perform the split. The "after_section" slot is determined by the next match to the regular expresion "^[#] .*$".}
}
\description{
This function takes the README.md in the working directory and converts it into a README class object
}
