% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_StandardFileDetailsDF.R
\name{make_StandardFileDetailsDF}
\alias{make_StandardFileDetailsDF}
\title{Create a File Details Dataframe}
\usage{
make_StandardFileDetailsDF(
  sources,
  output_file,
  summary,
  disable_rstudioapi = FALSE
)
}
\arguments{
\item{disable_rstudioapi}{TRUE if the RStudioAPI should be disabled. Errors can be thrown in Jobs and running scripts from the command line.}
}
\description{
Create a File Details Dataframe
}
