% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperclip.R
\name{paperclip}
\alias{paperclip}
\title{Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.}
\usage{
paperclip(
  destination_file,
  sources,
  output_file,
  summary,
  ...,
  disable_rstudioapi = FALSE
)
}
\description{
Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.
}
