% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_readi_class.R
\docType{class}
\name{Readi-class}
\alias{Readi-class}
\alias{Readi}
\title{Set Readi S4 Class
A Readi object contains the contents of a README File Details section and a dataframe on the same File Details pertaining to a specific Excel or csv file written out of R.}
\description{
Set Readi S4 Class
A Readi object contains the contents of a README File Details section and a dataframe on the same File Details pertaining to a specific Excel or csv file written out of R.
}
\section{Slots}{

\describe{
\item{\code{dataframe}}{dataframe of the Readi arguments specific to a given dataframe output file}

\item{\code{readme}}{full README read from the README in the working directory as a character vector.}
}}

