% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Classes.R
\docType{class}
\name{README-class}
\alias{README-class}
\alias{README}
\title{Set README S4 Class
A README object contains the contents of a README split by a given section name}
\description{
Set README S4 Class
A README object contains the contents of a README split by a given section name
}
\section{Slots}{

\describe{
\item{\code{section_headings}}{all the lines in the README.md input that begin with regular expression "^# .*$"}

\item{\code{before_section}}{vector of lines before a given section}

\item{\code{section}}{vector of lines within a given section}

\item{\code{after_section}}{vector of lines after a given section}
}}

