% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_metrics.R
\name{report_metrics}
\alias{report_metrics}
\title{Return metrics on all the Tabs in an Excel}
\usage{
report_metrics(path_to_excel)
}
\arguments{
\item{path_to_excel}{path to the Excel to generate metrics on}
}
\value{
dataframe of 1. Object Name, 2. Object Class, 3. Length, 4. Names, 5. Rows, 6. Cols, and 7. Column Names. Vectors that were originally of length 1 or greater were converted to a string that is parseable into an expression using the vector_to_string function in the cave package. The value can be converted back to a vector using the cave package's string_to_vector function.
}
\description{
Also note that this function is similar to the trace_objects() function, except the this function is meant to  report the metrics of the final data written to the tabs while the trace_objects() is tracing the provenance prior to writing to Excel.
}
