% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FileDetails.R
\docType{data}
\name{sources}
\alias{sources}
\title{Make a FileDetails class object}
\format{
An object of class \code{character} of length 1.
}
\usage{
sources
}
\description{
Make a FileDetails class object
}
\keyword{datasets}
