% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple_FileDetails_to.R
\name{staple_FileDetails_to}
\alias{staple_FileDetails_to}
\title{Attach FileDetails to the README.md and Destination File}
\usage{
staple_FileDetails_to(
  destination_file,
  sources,
  output_file,
  summary,
  ...,
  disable_rstudioapi = FALSE
)
}
\description{
As opposed to "add_" functions, "staple_" functions attach FileDetails to the README.md and destination file in the same function.
}
