% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_heading.R
\name{to_heading}
\alias{to_heading}
\title{Get a section_name in GitHub Flavored Markdown heading format}
\usage{
to_heading(section_name)
}
\description{
Get a section_name in GitHub Flavored Markdown heading format
}
