% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_objects.R
\name{trace_objects}
\alias{trace_objects}
\title{Return metadata on the objects in the global environment}
\usage{
trace_objects(objname_pattern)
}
\arguments{
\item{objname_pattern}{pattern argument for the ls() base function}
}
\value{
dataframe of 1. Object Name, 2. Object Class, 3. Length, 4. Names, 5. Rows, 6. Cols, and 7. Column Names. Vectors that were originally of length 1 or greater were converted to a string that is parseable into an expression using the vector_to_string function in the cave package. The value can be converted back to a vector using the cave package's string_to_vector function.
}
\description{
This function allows to view the series of transformations that occur in a given script. This function will only work effectively if the global environment only contains objects that are to be traced and all the object naming conventions are named such that the stem of the label match the pattern to trace the transformations in sequential order. For example an input can initially be input1 and after a transformation, the next version will be named input2 and the next input3 and so on. The complete list of teh glboal environment can also be obtained by not entering an objname_pattern.
}
