% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_excel.R
\name{update_excel}
\alias{update_excel}
\title{Update an Excel file with a File Details tab}
\usage{
update_excel(readi_object, path_to_excel, ...)
}
\description{
Update an Excel file with a File Details tab
}
