% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_readme.R
\name{update_readme}
\alias{update_readme}
\title{Update README.md in working directory from a Readi obj}
\usage{
update_readme(readi_object)
}
\description{
Update README.md in working directory from a Readi obj
}
