% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple_FileDetails_to.R
\name{staple_FileDetails_to}
\alias{staple_FileDetails_to}
\title{Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.}
\usage{
staple_FileDetails_to(
  destination_file,
  sources,
  output_file,
  summary,
  ...,
  disable_rstudioapi = FALSE
)
}
\description{
Applies a Readi to README.md File Details section with the new file details as well as attaches it to the provided csv or Excel file.
}
